Add files via upload
These files are for a webpage that allows the user to change the background with a single click. The user has two options, click the Hex where it will change the background into a random colors. Or the simple version where it changes the background to pre-selected colors.

Also has stuff I use for a template in general. You can ignore those.